Organize scope service helper tests
