+++ test/services/helpers/scopePermissions/ScopeService.test.js
