+++ test/services/helpers/scopePermissions/hooks.test.js
@@ -1,0 +1,8 @@
+/* eslint-disable global-require */
+/**
+ * This file is only used to group the hook tests together
+ */
+describe('ScopeService hooks', () => {
+	require('./hooks/lookupScope');
+	require('./hooks/rejectQueryingForOtherUsers');
+});
